requirements.txt
python-telegram-bot==20.7
openpyxl==3.1.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
